pyproject.toml
[project]
name = "echo"
version = "0.2.0"
description = "基于 FastAPI + Milvus + RAG 的智能客服后端系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4.0",
    "langchain-openai>=1.0.0",
    "langchain-milvus>=0.1.0",
    "langchain-text-splitters>=1.0.0",
    "pymilvus>=2.4.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pypdf>=6.2.0",
]
